flipzen/requirements.txt
-e git+https://github.com/VisionOra/NewsVerifier---Agent.git@4a58ee103dbbacbf82d253a4d89476581bc10f74#egg=agent
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
blockbuster==1.5.24
bs4==0.0.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
cryptography==44.0.3
cssselect==1.3.0
distro==1.9.0
fastapi==0.115.12
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
forbiddenfruit==0.1.4
greenlet==3.2.1
h11==0.16.0
html2text==2025.4.15
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jieba3k==0.35.1
jiter==0.9.0
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
langchain==0.3.25
langchain-core==0.3.58
langchain-text-splitters==0.3.8
langgraph==0.4.2
langgraph-api==0.2.15
langgraph-checkpoint==2.0.25
langgraph-cli==0.2.8
langgraph-prebuilt==0.1.8
langgraph-runtime-inmem==0.0.9
langgraph-sdk==0.1.66
langsmith==0.3.42
lxml==5.4.0
lxml_html_clean==0.4.2
newspaper3k==0.2.8
nltk==3.9.1
openai==1.77.0
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
pillow==11.2.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
sse-starlette==2.1.3
starlette==0.46.2
structlog==25.3.0
tenacity==9.1.2
tinysegmenter==0.3
tldextract==5.3.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
xxhash==3.5.0
zstandard==0.23.0
